requirements.txt
asn1crypto==0.24.0
astroid==1.5.3
attrs==17.4.0
Automat==0.6.0
autopep8==1.3.3
bcrypt==3.1.4
beautifulsoup4==4.6.0
breadability==0.1.20
bs4==0.0.1
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==2.1.4
docopt==0.6.2
Fabric3==1.13.1.post1
flake8==3.5.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
isort==4.2.15
jieba==0.39
lazy-object-proxy==1.3.1
lxml==4.1.1
mccabe==0.6.1
nltk==3.2.5
nose==1.3.7
paramiko==2.4.0
psutil==5.4.3
pyasn1==0.4.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyNaCl==1.2.1
requests==2.18.4
rope==0.10.7
six==1.11.0
sumy==0.7.0
tqdm==4.19.5
urllib3==1.22
wrapt==1.10.11
yapf==0.20.1
zope.interface==4.4.3
